+++ test/plugin/scenarios/resteasy-6.x-scenario/src/main/java/test/apache/skywalking/apm/testcase/UserResource.java
@@ -1,0 +1,78 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package test.apache.skywalking.apm.testcase;
+
+import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.DELETE;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.POST;
+import jakarta.ws.rs.PUT;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+import test.apache.skywalking.apm.testcase.entity.User;
+
+import java.net.URI;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+
+@Path("/")
+public class UserResource {
+
+    private static final Map<Integer, User> USERS = new ConcurrentHashMap<>();
+
+    @GET
+    @Path("/get/{id}")
+    @Produces(MediaType.APPLICATION_JSON)
+    public Response getUser(@PathParam("id") int id) {
+        User currentUser = new User(id, "a");
+        return Response.ok(currentUser).build();
+    }
+
+    @POST
+    @Path("/create/")
+    @Produces(MediaType.APPLICATION_JSON)
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response createUser(User user) {
+        USERS.put(user.getId(), user);
+        return Response.created(URI.create("")).build();
+    }
+
+    @PUT
+    @Path("/update/{id}")
+    @Produces(MediaType.APPLICATION_JSON)
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response updateUser(@PathParam("id") int id, User user) {
+        User currentUser = new User(id, user.getUserName());
+        return Response.ok(currentUser).build();
+    }
+
+    @DELETE
+    @Path("/delete/{id}")
+    public Response deleteUser(@PathParam("id") int id) {
+        User currentUser = USERS.get(id);
+        if (currentUser == null) {
+            return Response.noContent().build();
+        }
+        USERS.remove(id);
+        return Response.noContent().build();
+    }
+}